+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/services/configurers/implementations/DefaultFileConfigurer.java
@@ -120,7 +120,6 @@ public class DefaultFileConfigurer implements FileConfigurer<DefaultFileWrapper>
         userPermissions.setFillForms(
                 !action.equals(Action.view)
                         && !action.equals(Action.comment)
-                        && !action.equals(Action.embedded)
                         && !action.equals(Action.blockcontent)
         );